server/requirements.txt
# server/requirements.txt — Dependencies for Docker build
# Mirrors echo_env/server/requirements.txt pattern

openenv-core[core]>=0.2.2
fastapi>=0.115.0
pydantic>=2.0.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
fastmcp>=2.0.0
openai>=1.12.0
python-dotenv>=1.0.0
httpx>=0.25.0
